+++ lib/src/segyio/util.h
@@ -19,7 +19,7 @@ void ascii2ebcdic( const char* ascii, char* ebcdic );
 void ibm2ieee(void* to, const void* from);
 void ieee2ibm(void* to, const void* from);
 int segy_seek( struct segy_file_handle*, int, long, unsigned int );
-long segy_ftell( struct segy_file_handle* );
+long long segy_ftell( struct segy_file_handle* );
 
 #ifdef __cplusplus
 }
